Removed duplicated utils, reorganizing utils
